+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Show progress of Chat request/question processing
 - Add Goal and Task Details in Navigation Bar to view Description, Acceptance Criteria and Status
 - Improvements to Chat Response UI and Code Block rendering
+- Enable Clipboard Copy of Sara Responses
 
 ### Bug Fixes